BUFR Table cleanup; add 17, 18, 19; use 19.1.1 as default tables

+++ grib/src/main/java/ucar/nc2/grib/grib2/Grib2Gds.java
@@ -1029,7 +1029,7 @@ Template 3.40 (Grid definition template 3.40 - Gaussian latitude/longitude)
 
     public GdsHorizCoordSys makeHorizCoordSys() {
 
-      int nlats = (int) (2 * Nparellels);
+      int nlats = 2 * Nparellels;
       GaussianLatitudes gaussLats = new GaussianLatitudes(nlats);
 
       int bestStartIndex = 0, bestEndIndex = 0;
